refactor: конструктор бургера теперь состоит из OrderConstructor и BurgerConstructor
