+++ src/components/burger-constructor/burger-constructor.module.css
@@ -17,9 +17,3 @@
   display: flex;
   align-items: center;
 }
-
-.footer {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-}
