+++ src/components/order-constructor/order-constructor.module.css
@@ -1,0 +1,5 @@
+.footer {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+}
